requirement.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-groq
langchain-core
langchain-community
langgraph